+++ packages/atomic/cypress/integration/facets-v1/facet/facet-assertions.ts
@@ -150,6 +150,15 @@ export function assertDisplaySearchClearButton(display: boolean) {
   });
 }
 
+export function assertHighlightsResults(query: string) {
+  it(`should highlight the results with the query "${query}"`, () => {
+    FacetSelectors.valueHighlight().each((element) => {
+      const text = element.text().toLowerCase();
+      expect(text).to.eq(query.toLowerCase());
+    });
+  });
+}
+
 export function assertSearchInputEmpty() {
   it('the search input should be empty', () => {
     FacetSelectors.searchInput().invoke('val').should('be.empty');