+++ packages/atomic/src/components/facets-v1/atomic-timeframe-facet/atomic-timeframe-facet.tsx
@@ -1,0 +1,170 @@
+import {Component, h, State, Prop, VNode, Host} from '@stencil/core';
+import {
+  SearchStatus,
+  SearchStatusState,
+  buildSearchStatus,
+  DateFacet,
+  DateFacetState,
+  buildDateFacet,
+  DateFacetOptions,
+  DateFacetValue,
+} from '@coveo/headless';
+import {
+  Bindings,
+  BindStateToController,
+  InitializableComponent,
+  InitializeBindings,
+} from '../../../utils/initialization-utils';
+import {FacetPlaceholder} from '../../facets/atomic-facet-placeholder/atomic-facet-placeholder';
+import {FacetContainer} from '../facet-container/facet-container';
+import {FacetHeader} from '../facet-header/facet-header';
+import {FacetValueLink} from '../facet-value-link/facet-value-link';
+import {BaseFacet} from '../facet-common';
+
+/**
+ * A facet is a list of values for a certain field occurring in the results.
+ * An `atomic-timeframe-facet` displays a facet of the results for the current query as date intervals.
+ *
+ * @part facet - The wrapper for the entire facet.
+ * @part placeholder - The placeholder shown before the first search is executed.
+ *
+ * @part label-button - The button that displays the label and allows to expand/collapse the facet.
+ * @part label-button-icon - The label button icon.
+ * @part clear-button - The button that resets the actively selected facet values.
+ * @part clear-button-icon - The clear button icon.
+ *
+ * @part values - The facet values container.
+ * @part value-label - The facet value label, common for all displays.
+ * @part value-count - The facet value count, common for all displays.
+ * @part value-link - The facet value when display is 'link'.
+ */
+@Component({
+  tag: 'atomic-timeframe-facet',
+  styleUrl: 'atomic-timeframe-facet.pcss',
+  shadow: true,
+})
+export class AtomicTimeframeFacet
+  implements InitializableComponent, BaseFacet<DateFacet, DateFacetState> {
+  @InitializeBindings() public bindings!: Bindings;
+  public facet!: DateFacet;
+  public searchStatus!: SearchStatus;
+
+  @BindStateToController('facet')
+  @State()
+  public facetState!: DateFacetState;
+  @BindStateToController('searchStatus')
+  @State()
+  public searchStatusState!: SearchStatusState;
+  @State() public error!: Error;
+  @State() public isCollapsed = false;
+
+  /**
+   * Specifies a unique identifier for the facet.
+   */
+  @Prop({mutable: true, reflect: true}) public facetId?: string;
+  /**
+   * The non-localized label for the facet.
+   */
+  @Prop() public label = 'no-label';
+  /**
+   * The field whose values you want to display in the facet.
+   */
+  @Prop() public field = 'date';
+  /**
+   * Whether this facet should contain an datepicker allowing users to set custom ranges.
+   */
+  @Prop() public withDatePicker = false;
+
+  public initialize() {
+    this.searchStatus = buildSearchStatus(this.bindings.engine);
+    const options: DateFacetOptions = {
+      facetId: this.facetId,
+      field: this.field,
+      currentValues: [], // TODO: parse ranges
+      generateAutomaticRanges: false,
+    };
+    this.facet = buildDateFacet(this.bindings.engine, {options});
+    this.facetId = this.facet.state.facetId;
+    this.bindings.store.state.facets[this.facetId] = {
+      label: this.label,
+    };
+  }
+
+  private get numberOfSelectedValues() {
+    return this.facetState.values.filter(({state}) => state === 'selected')
+      .length;
+  }
+
+  private renderHeader() {
+    return (
+      <FacetHeader
+        i18n={this.bindings.i18n}
+        label={this.label}
+        onClearFilters={() => this.facet.deselectAll()}
+        numberOfSelectedValues={this.numberOfSelectedValues}
+        isCollapsed={this.isCollapsed}
+        onToggleCollapse={() => (this.isCollapsed = !this.isCollapsed)}
+      ></FacetHeader>
+    );
+  }
+
+  private renderValue(facetValue: DateFacetValue) {
+    return (
+      <FacetValueLink
+        displayValue={`${facetValue.start} - ${facetValue.end}`} // TODO: format display value
+        numberOfResults={facetValue.numberOfResults}
+        isSelected={facetValue.state === 'selected'}
+        i18n={this.bindings.i18n}
+        onClick={() => this.facet.toggleSingleSelect(facetValue)}
+      ></FacetValueLink>
+    );
+  }
+
+  private renderValuesContainer(children: VNode[]) {
+    return (
+      <ul part="values" class="mt-3">
+        {children}
+      </ul>
+    );
+  }
+
+  private renderValues() {
+    return this.renderValuesContainer(
+      this.valuesToRender.map((value) => this.renderValue(value))
+    );
+  }
+
+  private get valuesToRender() {
+    return this.facetState.values.filter(
+      (value) => value.numberOfResults || value.state !== 'idle'
+    );
+  }
+
+  public render() {
+    if (this.searchStatusState.hasError) {
+      return;
+    }
+
+    if (!this.searchStatusState.firstSearchExecuted) {
+      return (
+        <FacetPlaceholder
+          numberOfValues={6} // TODO: use number of ranges
+        ></FacetPlaceholder>
+      );
+    }
+
+    if (!this.valuesToRender.length) {
+      return <Host class="atomic-without-values"></Host>;
+    }
+
+    return (
+      <Host class="atomic-with-values">
+        <FacetContainer>
+          {this.renderHeader()}
+          {!this.isCollapsed && this.renderValues()}
+          {/* TODO: add date picker */}
+        </FacetContainer>
+      </Host>
+    );
+  }
+}