+++ packages/headless/src/controllers/facets/range-facet/numeric-facet/headless-numeric-filter.test.ts
@@ -21,7 +21,6 @@ import {
 import {buildMockNumericFacetValue} from '../../../../test/mock-numeric-facet-value';
 import {executeSearch} from '../../../../features/search/search-actions';
 import {updateFacetOptions} from '../../../../features/facet-options/facet-options-actions';
-import {deselectAllFacetValues} from '../../../../features/facets/facet-set/facet-set-actions';
 import {buildMockNumericFacetResponse} from '../../../../test/mock-numeric-facet-response';
 
 describe('numeric filter', () => {
@@ -129,13 +128,25 @@ describe('numeric filter', () => {
       );
       expect(action).toBeTruthy();
     });
+
+    it('should return true when range is valid', () => {
+      const value = buildMockNumericFacetValue({start: 5, end: 10});
+      expect(numericFacet.setRange(value)).toBe(true);
+    });
+
+    it('should return false when range start value is greater than range end value', () => {
+      const value = buildMockNumericFacetValue({start: 10, end: 5});
+      expect(numericFacet.setRange(value)).toBe(false);
+    });
   });
 
   describe('#clear', () => {
     beforeEach(() => numericFacet.clear());
 
-    it('dispatches #deselectAllFacetValues with the facet id', () => {
-      expect(engine.actions).toContainEqual(deselectAllFacetValues(facetId));
+    it('dispatches #updateNumericFacetValues with the facet id and an empty array', () => {
+      expect(engine.actions).toContainEqual(
+        updateNumericFacetValues({facetId, values: []})
+      );
     });
 
     it('dispatches a #updateFacetOptions action with #freezeFacetOrder true', () => {
@@ -153,13 +164,22 @@ describe('numeric filter', () => {
     });
   });
 
-  it('exposes a #state getter property to retrieve the values', () => {
-    const value = buildMockNumericFacetValue();
+  it('the state #range property should return the range if it is selected', () => {
+    const value = buildMockNumericFacetValue({state: 'selected'});
     state.search.response.facets = [
       buildMockNumericFacetResponse({facetId, values: [value]}),
     ];
 
     expect(numericFacet.state.range).toEqual(value);
+  });
+
+  it('the state #range property should not return the range if it is not selected', () => {
+    const value = buildMockNumericFacetValue({state: 'idle'});
+    state.search.response.facets = [
+      buildMockNumericFacetResponse({facetId, values: [value]}),
+    ];
+
+    expect(numericFacet.state.range).toBeUndefined();
   });
 
   it('#state.facetId exposes the facet id', () => {